Move tree manager tests to integration tests
They're really integration tests and we should do more minimal unit
testing for the tree manager instead.
